Unite all Create CSV File functions (#663)
