+++ artifactory/commands/utils/transfer.go
@@ -3,10 +3,13 @@ package utils
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/gocarina/gocsv"
+	logutils "github.com/jfrog/jfrog-cli-core/v2/utils/log"
 	"github.com/jfrog/jfrog-client-go/http/jfroghttpclient"
 	"github.com/jfrog/jfrog-client-go/utils/errorutils"
 	"github.com/jfrog/jfrog-client-go/utils/io/httputils"
 	"net/http"
+	"time"
 )
 
 type ServerType string
@@ -42,4 +45,22 @@ func GetTransferPluginVersion(client *jfroghttpclient.JfrogHttpClient, url, plug
 
 type VersionResponse struct {
 	Version string `json:"version,omitempty"`
+}
+
+func CreateCSVFile(filePrefix string, items interface{}, timeStarted time.Time) (csvPath string, err error) {
+	// Create CSV file
+	summaryCsv, err := logutils.CreateCustomLogFile(fmt.Sprintf("%s-%s.csv", filePrefix, timeStarted.Format(logutils.DefaultLogTimeLayout)))
+	if err != nil {
+		return
+	}
+	csvPath = summaryCsv.Name()
+	defer func() {
+		e := summaryCsv.Close()
+		if err == nil {
+			err = e
+		}
+	}()
+	// Marshal JSON typed items array to CSV file
+	err = errorutils.CheckError(gocsv.MarshalFile(items, summaryCsv))
+	return
 }
